requirements.txt
discord.py>=2.0.1
pypika
pydantic==2.*
cachetools
python-dateutil
sortedcontainers
hllrcon==0.2.*
